# db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.25.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
	"github.com/ugent-library/projects-service/models"
)

type Project struct {
	ID              int64
	Name            []models.Translation
	Description     []models.Translation
	FoundingDate    pgtype.Text
	DissolutionDate pgtype.Text
	Attributes      []models.Attribute
	CreatedAt       pgtype.Timestamptz
	UpdatedAt       pgtype.Timestamptz
}

type ProjectsIdentifier struct {
	ProjectID int64
	Type      string
	Value     string
}